Refactor code, move similiar code from Interface- and TypeCreator to common superclass
