+++ common/schema-builder/src/main/java/io/smallrye/graphql/schema/SchemaBuilder.java
@@ -141,6 +141,7 @@ public class SchemaBuilder {
 
     private void setupDirectives(Directives directives) {
         typeCreator.setDirectives(directives);
+        interfaceCreator.setDirectives(directives);
         fieldCreator.setDirectives(directives);
     }
